Correccion de errores en CU Cobrar pedido

+++ ItaliaPizza/DataAccess/supplierIngredient.cs
@@ -17,7 +17,7 @@ namespace DataAccess
         public int idSupplierIngredient { get; set; }
         public string idSupplierOrder { get; set; }
         public int idIngredient { get; set; }
-        public double quantity { get; set; }
+        public int quantity { get; set; }
     
         public virtual ingredient ingredient { get; set; }
         public virtual supplierOrder supplierOrder { get; set; }
